refactor: simplify ipc client builder
