+++ src/electron/requests/files/renderer.ts
@@ -1,19 +1,6 @@
-import { ipcRenderer } from 'electron';
+import { ipcRendererFetcher } from '../../utils/ipc/ipcRendererFetcher';
 
 import { filesChannel } from '.';
 
 export const { importNotes, selectDirectory, getUserDataPath, getResourcesPath } =
-	filesChannel.client({
-		async importNotes({ channelName }) {
-			return ipcRenderer.invoke(channelName);
-		},
-		async selectDirectory({ channelName }) {
-			return ipcRenderer.invoke(channelName);
-		},
-		async getUserDataPath({ channelName, args }) {
-			return ipcRenderer.invoke(channelName, args);
-		},
-		async getResourcesPath({ channelName, args }) {
-			return ipcRenderer.invoke(channelName, args);
-		},
-	});
+	filesChannel.client(ipcRendererFetcher);
