+++ src/electron/requests/interactions/main.ts
@@ -1,6 +1,6 @@
 import { shell } from 'electron';
 
-import { ipcMainHandler } from '../../utils/ipc/electronMain';
+import { ipcMainHandler } from '../../utils/ipc/ipcMainHandler';
 
 import { interactionsChannel } from '.';
