+++ src/electron/utils/ipc/index.ts
@@ -2,11 +2,11 @@ export interface ChannelOptions {
 	readonly name: string;
 }
 
-export type ApiToFetchers<T extends Record<string, (...args: any[]) => Promise<any>>> = {
-	[K in keyof T]: (parameters: {
-		channelName: string;
-		args: Parameters<T[K]>;
-	}) => ReturnType<T[K]>;
+export type ApiToMappers<T extends Record<string, (...args: any[]) => Promise<any>>> = {
+	[K in keyof T]: (
+		rsp: T[K] extends Promise<infer R> ? R : never,
+		args: Parameters<T[K]>,
+	) => ReturnType<T[K]>;
 };
 
 export type ApiToHandlers<
@@ -24,22 +24,23 @@ export type ServerRequestHandler<Context = never> = (
 	callback: (parameters: { req: any; ctx: Context }) => any,
 ) => () => void;
 
+export type ClientFetcher = (endpoint: string, args: any[]) => Promise<any>;
+
 export const createChannel = <T extends Record<string, (...args: any[]) => Promise<any>>>(
 	options: ChannelOptions,
 ) => {
 	return {
-		// TODO: return proxy object that will forward calls to a fetcher with optional data mapping
-		client: (fetchers: ApiToFetchers<T>): T => {
-			return Object.fromEntries(
-				Object.entries(fetchers).map(([methodName, callback]) => [
-					methodName,
-					(...args) =>
-						callback({
-							channelName: [options.name, methodName].join('.'),
-							args,
-						}),
-				]),
-			) as T;
+		client: (fetcher: ClientFetcher, mappers: Partial<ApiToMappers<T>> = {}): T => {
+			return new Proxy<T>({} as any, {
+				get: function (_target, methodName: string) {
+					const endpoint = [options.name, methodName].join('.');
+					const mapper = mappers[methodName];
+					return (...args: Parameters<T[keyof T]>) =>
+						fetcher(endpoint, args).then((response) =>
+							mapper ? mapper(response, args) : response,
+						);
+				},
+			});
 		},
 		server: <Context = never>(
 			requestsHandler: ServerRequestHandler<Context>,
