+++ src/electron/utils/ipc/ipcRendererFetcher.ts
@@ -1,0 +1,4 @@
+import { ipcRenderer } from 'electron';
+
+export const ipcRendererFetcher = (channelName: string, args: any[]) =>
+	ipcRenderer.invoke(channelName, args);
